requirements.txt
flask[async]
pydantic
aiosqlite
flake8
pytest
pytest-asyncio
streamlit
fastapi
uvicorn
httpx
redis
black
pyjwt